# src/frontend/openapi/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "ResourcePulse.Hosting",
    "version": "1.0"
  },
  "paths": {
...
    "/api/projects/{id}": {
      "put": {
        "tags": [
          "Projects"
        ],
        "operationId": "Projects_UpdateProject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/projects/{id}/start": {
      "post": {
        "tags": [
          "Projects"
        ],
        "operationId": "Projects_Start",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/projects/{id}/complete": {
      "post": {
        "tags": [
          "Projects"
        ],
        "operationId": "Projects_Complete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/projects/{id}/suspend": {
      "post": {
        "tags": [
          "Projects"
        ],
        "operationId": "Projects_Suspend",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/projects/{id}/resume": {
      "post": {
        "tags": [
          "Projects"
        ],
        "operationId": "Projects_Resume",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/projects/{id}/cancel": {
      "post": {
        "tags": [
          "Projects"
        ],
        "operationId": "Projects_Cancel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/projects/{id}/skill-requirements": {
      "post": {
        "tags": [
          "Projects"
        ],
        "operationId": "Projects_AddSkillRequirement",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddOrUpdateProjectSkillRequirementDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddOrUpdateProjectSkillRequirementDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddOrUpdateProjectSkillRequirementDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/projects/{id}/skill-requirements/{skillId}": {
      "put": {
        "tags": [
          "Projects"
        ],
        "operationId": "Projects_UpdateSkillRequirement",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "skillId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddOrUpdateProjectSkillRequirementDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddOrUpdateProjectSkillRequirementDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddOrUpdateProjectSkillRequirementDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "operationId": "Projects_RemoveSkillRequirement",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "skillId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/projects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "operationId": "Projects_GetActiveInRange",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "dateSource",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/DateSource"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectNodeReadDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectNodeReadDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectNodeReadDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddOrUpdateProjectSkillRequirementDto": {
        "type": "object",
        "properties": {
          "skillId": {
            "type": "string",
            "format": "uuid"
          },
          "minLevel": {
            "$ref": "#/components/schemas/SkillLevel"
          }
        },
        "additionalProperties": false
      },
...
      "CommitmentLevel": {
        "enum": [
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32",
        "x-enum-varnames": [
          "Exploratory",
          "Planned",
          "Committed",
          "Critical"
        ]
      },
...
      "DateSource": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32",
        "x-enum-varnames": [
          "Planned",
          "Baseline",
          "Effective"
        ]
      },
...
      "DerivedExecutionStatus": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32",
        "x-enum-varnames": [
          "NotStarted",
          "InProgress",
          "Completed"
        ]
      },
...
      "PlanningMode": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32",
        "x-enum-varnames": [
          "Unspecified",
          "FixedWork",
          "FixedDuration"
        ]
      },
...
      "ProjectNodeReadDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "nodeType": {
            "$ref": "#/components/schemas/ProjectNodeType"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "depth": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "$ref": "#/components/schemas/ProjectType"
          },
          "commitmentLevel": {
            "$ref": "#/components/schemas/CommitmentLevel"
          },
          "leadResourceId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "leadResourceName": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "client": {
            "type": "string",
            "nullable": true
          },
          "baselineStart": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "baselineEnd": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "baselinedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "plannedStart": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "plannedEnd": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "actualStart": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "actualEnd": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "planningMode": {
            "$ref": "#/components/schemas/PlanningMode"
          },
          "estimatedWork": {
            "type": "string",
            "format": "date-span",
            "nullable": true
          },
          "scheduleVarianceStart": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "scheduleVarianceEnd": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "forecastVarianceEnd": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isLate": {
            "type": "boolean",
            "nullable": true
          },
          "derivedStatus": {
            "$ref": "#/components/schemas/DerivedExecutionStatus"
          },
          "isProposed": {
            "type": "boolean",
            "nullable": true
          },
          "skillRequirements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectSkillRequirementDto"
            },
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectNodeTagDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProjectNodeTagDto": {
        "type": "object",
        "properties": {
          "tagId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "ProjectNodeType": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32",
        "x-enum-varnames": [
          "Project",
          "Phase",
          "WorkPackage"
        ]
      },
      "ProjectSkillRequirementDto": {
        "type": "object",
        "properties": {
          "skillId": {
            "type": "string",
            "format": "uuid"
          },
          "minLevel": {
            "$ref": "#/components/schemas/SkillLevel"
          }
        },
        "additionalProperties": false
      },
      "ProjectStatus": {
        "enum": [
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32",
        "x-enum-varnames": [
          "Draft",
          "Active",
          "OnHold",
          "Closed",
          "Cancelled"
        ]
      },
      "ProjectType": {
        "enum": [
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32",
        "x-enum-varnames": [
          "Internal",
          "Customer",
          "Investment",
          "Maintenance"
        ]
      },
      "ReasonDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SkillLevel": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32",
        "x-enum-varnames": [
          "Basic",
          "Proficient",
          "Expert"
        ]
      },
...
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/ProjectType"
          },
          "commitmentLevel": {
            "$ref": "#/components/schemas/CommitmentLevel"
          },
          "leadResourceId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "client": {
            "type": "string",
            "nullable": true
          },
          "confirmDemoteHardAllocations": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}